+++ plugins/vmpayment/emspayklarnapaylater/emspayklarnapaylater.php
@@ -77,7 +77,7 @@ class plgVmPaymentEmspayklarnaPayLater extends EmspayVmPaymentPlugin
                 $methodSalesPrice = $this->setCartPrices($cart, $cartPrices, $this->_currentMethod);
                 $this->_currentMethod->$method_name = $this->renderPluginName($this->_currentMethod);
                 $html = $this->getPluginHtml($this->_currentMethod, $selected, $methodSalesPrice);
-                $htmla[] = $html . '<br />' . $this->customInfoHTML();
+                $htmla[] = $html . '<br />';
             }
         }
         $htmlIn[] = $htmla;
@@ -99,31 +99,6 @@ class plgVmPaymentEmspayklarnaPayLater extends EmspayVmPaymentPlugin
         }
         return false;
     }
-
-    /**
-     * generate html for the gender selectbox
-     *
-     * @return string
-     * @since v1.0.0
-     */
-    public function customInfoHTML()
-    {
-        $html = JText::_('PLG_VMPAYMENT_EMSPAYKLARNAPAYLATER_MESSAGE_SELECT_GENDER') . ' <br/>';
-        $html .= '<select name="gender" id="' . $this->name . '" class="' . $this->name . '">';
-        $html .= '<option value="male" '
-                . (JFactory::getSession()->get('emspayklarnapaylater_gender') == 'male' ? " selected" : "") . '>'
-                . JText::_('PLG_VMPAYMENT_EMSPAYKLARNAPAYLATER_MESSAGE_SELECT_GENDER_MALE') . '</option>';
-        $html .= '<option value="female" '
-                . (JFactory::getSession()->get('emspayklarnapaylater_gender') == 'male' ? " selected" : "") . '>'
-                . JText::_('PLG_VMPAYMENT_EMSPAYKLARNAPAYLATER_MESSAGE_SELECT_GENDER_FEMALE') . '</option>';
-        $html .= "</select><br/>";
-        $html .= JText::_('PLG_VMPAYMENT_EMSPAYKLARNAPAYLATER_MESSAGE_ENTER_DOB') . '<br>';
-        $html .= '<input type="text" name="dob" value="' . JFactory::getSession()->get('emspayklarnapaylater_dob', null, 'vm') . '"/>';
-        $html .= '<i>('.JText::_("PLG_VMPAYMENT_EMSPAYKLARNAPAYLATER_MESSAGE_DATE_FORMAT").')</i>';
-        
-        return $html;
-    }
-
             
     /**
      * This is for checking the input data of the payment method within the checkout
@@ -139,15 +114,6 @@ class plgVmPaymentEmspayklarnaPayLater extends EmspayVmPaymentPlugin
         }
 
         if (!($this->_currentMethod = $this->getVmPluginMethod($cart->virtuemart_paymentmethod_id))) {
-            return false;
-        }
-        
-        $app = JFactory::getApplication();
-        $dob = $app->getSession()->get('emspayklarnapaylater_dob', null, 'vm');
-        if ($this->isValidDate($dob) === false) {
-            $app->enqueueMessage(JText::_("PLG_VMPAYMENT_EMSPAYKLARNAPAYLATER_MESSAGE_INVALID_DATE_ERROR"), 'error');
-            $app->getSession()->clear('emspayklarnapaylater_dob', 'vm');
-            $app->redirect(JRoute::_('index.php?option=com_virtuemart&view=cart&task=editpayment', false));
             return false;
         }
         return true;
@@ -171,25 +137,7 @@ class plgVmPaymentEmspayklarnaPayLater extends EmspayVmPaymentPlugin
         if (!($currentMethod = $this->getVmPluginMethod($cart->virtuemart_paymentmethod_id))) {
             return false;
         }
-        JFactory::getSession()->set('emspayklarnapaylater_gender', vRequest::getVar('gender'), 'vm');
-        JFactory::getSession()->set('emspayklarnapaylater_dob', vRequest::getVar('dob'), 'vm');
-
         return true;
-    }
-
-    /**
-     * validate date of birth
-     *
-     * @param type $string
-     * @return boolean
-     * @since v1.0.0
-     */
-    protected function isValidDate($string)
-    {
-        if (preg_match('/^(\d{2})-(\d{2})-(\d{4})$/', $string, $matches)) {
-            return DateTime::createFromFormat('d-m-Y', $string) instanceof \DateTime;
-        }
-        return false;
     }
 
     /**
@@ -259,8 +207,6 @@ class plgVmPaymentEmspayklarnaPayLater extends EmspayVmPaymentPlugin
                         $order['details']['BT'],
                         \EmspayHelper::getLocale(),
                         filter_var(\JFactory::getApplication()->input->server->get('REMOTE_ADDR'), FILTER_VALIDATE_IP),
-                        \JFactory::getSession()->get('emspayklarnapaylater_gender', null, 'vm'),
-                        $this->convertDateToKlarnaPayLaterFormat(\JFactory::getSession()->get('emspayklarnapaylater_dob', null, 'vm'))
         );
 
         $plugin = ['plugin' => EmspayHelper::getPluginVersion($this->_name)];
@@ -404,18 +350,6 @@ class plgVmPaymentEmspayklarnaPayLater extends EmspayVmPaymentPlugin
         return round(0, 2);
     }
 
-    /**
-     * clear user seesion data
-     * @since v1.0.0
-     */
-    protected function clearSessionData()
-    {
-        $session = JFactory::getSession();
-        $session->clear('emspayklarnapaylater_gender', 'vm');
-        $session->clear('emspayklarnapaylater_dob', 'vm');
-    }
-
-    
     /**
      * on order status shipped
      *