+++ plugins/vmpayment/emspayklarnapaynow/emspayklarnapaynow.php
@@ -86,7 +86,7 @@ class plgVmPaymentEmspayKlarnaPayNow extends EmspayVmPaymentPlugin
                         \EmspayHelper::getLocale(),
                         filter_var(\JFactory::getApplication()->input->server->get('REMOTE_ADDR'), FILTER_VALIDATE_IP)
         );
-        $plugin = ['plugin' => EmspayHelper::getPluginVersion($this->_name)];
+        $plugin = ['plugin' => EmspayHelper::getPluginVersion()];
         $webhook =$this->getWebhookUrl(intval($order['details']['BT']->virtuemart_paymentmethod_id));
         
         try {