+++ plugins/vmpayment/emspaypaypal/emspaypaypal.php
@@ -87,7 +87,7 @@ class plgVmPaymentEmspaypaypal extends EmspayVmPaymentPlugin
                         \EmspayHelper::getLocale(),
                         filter_var(\JFactory::getApplication()->input->server->get('REMOTE_ADDR'), FILTER_VALIDATE_IP)
         );
-        $plugin = ['plugin' => EmspayHelper::getPluginVersion($this->_name)];
+        $plugin = ['plugin' => EmspayHelper::getPluginVersion()];
         $webhook =$this->getWebhookUrl(intval($order['details']['BT']->virtuemart_paymentmethod_id));
 
         try {