+++ plugins/vmpayment/emspaybancontact/emspaybancontact.php
@@ -17,7 +17,7 @@ use Emspay\Lib\EmspayVmPaymentPlugin;
  * @category    Ginger
  * @package     Ginger Virtuemart
  * @author      Ginger Payments B.V. (plugins@gingerpayments.com)
- * @version     v1.2.1
+ * @version     v1.3.0
  * @copyright   COPYRIGHT (C) 2018 GINGER PAYMENTS B.V.
  * @license     The MIT License (MIT)
  * @since       v1.0.0
@@ -29,7 +29,7 @@ if (!class_exists('vmPSPlugin')) {
 }
 
 JLoader::registerNamespace('Emspay', JPATH_LIBRARIES . '/emspay');
-JImport('emspay.ginger-php.vendor.autoload');
+JImport('emspay.vendor.autoload');
 JImport('emspay.emspayhelper');
 
 class plgVmPaymentEmspaybancontact extends EmspayVmPaymentPlugin